Refactor Close-RDCMan
Attempt to close via CloseMainWindow() initially, then (if Force specified) stop any remaining processes.

+++ RDCMan/Public/Close-RDCMan.ps1
@@ -7,14 +7,13 @@ function Close-RDCMan {
     $Running = Get-Process -Name RDCMan -ErrorAction SilentlyContinue
 
     foreach ($Process in $Running) {
-        if ($Force) {
-            Write-Verbose "[Close-RDCMan] Stopping RDCMan process [$($Process.Id) with title $($Process.MainWindowTitle) ..."
-            Stop-Process -Id $Process.Id -Force
-        }
+        $null = $Process.CloseMainWindow()
 
-        else {
-            Write-Verbose "[Close-RDCMan] Closing RDCMan process [$($Process.Id) with title $($Process.MainWindowTitle)"
-            $null = $Process.CloseMainWindow()
+        # Forcefully close remaining processes, if specified
+        if ($Force -and (Get-Process -Id $Process.Id -ErrorAction SilentlyContinue)) {
+            Write-Verbose "Process [$($Process.Id)] with title $($Process.MainWindowTitle) did not close gracefully, stopping [$($Process.Id)"
+
+            Stop-Process -Id $Process.Id -Force
         }
     }
 }
